requirements.txt
cv2=4.5.1
easygui=0.98.2
Pillow=8.2.0
PyAutoGUI=0.9.52
numpy==1.8.0


